Remove test_api_info.py
